Hw04 auth

+++ model/contacts.js
@@ -1,38 +1,48 @@
 const Contact = require('./schemas/contact')
 
-const listContacts = async () => {
-  const results = await Contact.find({})
-  return results
+const listContacts = async (userId,
+  { sortBy, sortByDesc, sub, limit = '20', page = '1' },
+) => {
+  const options = sub ? { owner: userId, subscription: `${sub}` } : { owner: userId }
+  const results = await Contact.paginate(options, {
+    limit,
+    page,
+    sort: {
+      ...(sortBy ? { [`${sortBy}`]: 1 } : {}), ...(sortByDesc ? { [`${sortByDesc}`]: -1 } : {})
+    },
+    populate: {
+      path: 'owner',
+      select: 'email -_id'
+    }
+  })
+  const { docs: contacts, totalDocs: total } = results
+  return { total: total.toString(), limit, page, contacts }
 }
 
-const getContactById = async (contactId) => {
-  const result = await Contact.findOne({ _id: contactId })
+const getContactById = async (contactId, userId) => {
+  const result = await Contact.findOne({ _id: contactId, owner: userId }).populate({
+    path: 'owner',
+    select: 'email -_id'
+  })
   return result
 }
 
 const addContact = async (body) => {
-  const contactsList = await listContacts()
-  const isExist = contactsList.find(({ name }) => name === body.name)
-  if (!isExist) {
-    const result = Contact.create(body)
-    return result
-  } else {
-    const message = `${body.name} is already in your contacts list`
-    return message
-  }
+  const result = await Contact.create(body)
+  return result
 }
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (contactId, body, userId) => {
   const result = await Contact.findByIdAndUpdate(
-    { _id: contactId },
+    { _id: contactId, owner: userId },
     { ...body },
     { new: true }
   )
   return result
 }
 
-const removeContact = async (contactId) => {
-  const result = await Contact.findOneAndRemove({ _id: contactId })
+const removeContact = async (contactId, userId) => {
+  const result = await Contact.findOneAndRemove({ _id: contactId, owner: userId })
   return result
 }
